User CRUD

+++ api/user/routes.py
@@ -2,8 +2,9 @@ from flask import Blueprint, jsonify, request
 from webargs import fields, ValidationError, flaskparser
 from flasgger import swag_from
 from webargs.flaskparser import use_args
-from api.user.models import User
+from api.user.models import User, Address
 from api.exceptions import UserExceptions
+from datetime import datetime
 
 from api.user.schemas import users_schema, user_schema
 
@@ -50,23 +51,25 @@ def get_users():
 
 @user_blueprint.route('/users/<user_id>', methods=['PUT', 'PATCH'])
 def update_user(user_id):
-    user = User.query.filter_by(id=user_id)
+    user = User.query.filter_by(id=user_id).first()
     data = request.json
     if not user:
         raise UserExceptions.user_not_exist()
-    user.update(data)
-    return jsonify(users_schema.dump(user)), HTTPStatus.OK
+    user.update(updated_at=datetime.utcnow(), **data)
+    return jsonify(user_schema.dump(user)), HTTPStatus.OK
 
 
 @user_blueprint.route('/users', methods=['POST'])
 def create_user():
-    data = request.body
+    data = request.json
     user = User.create(**data)
-    return user, HTTPStatus.OK
+    return jsonify(user_schema.dump(user)), HTTPStatus.OK
 
 
 @user_blueprint.route('/users/<user_id>', methods=['DELETE'])
 def delete_user(user_id):
-    user = User.query.filter_by(id=user_id)
+    user = User.query.filter_by(id=user_id).first()
+    if not user:
+        return '', HTTPStatus.NOT_FOUND
     user.delete()
-    return user, HTTPStatus.NO_CONTENT
+    return '', HTTPStatus.NO_CONTENT
